Added Popular events over, still needs some cleaning up to make it look nice
